Repo Gardening: move utilities into their own directory
As we've kept adding more and more small utility functions to `src/`, it has become quite messy. This should clean things up a bit.
