+++ projects/github-actions/repo-gardening/changelog/update-repo-gardening-utils-dir
@@ -1,0 +1,4 @@
+Significance: minor
+Type: changed
+
+General: move all utilities into their own directory to keep things ordered
