+++ projects/github-actions/repo-gardening/src/index.js
@@ -1,8 +1,5 @@
 const { setFailed, getInput } = require( '@actions/core' );
 const { context, getOctokit } = require( '@actions/github' );
-const debug = require( './debug' );
-const ifNotClosed = require( './if-not-closed' );
-const ifNotFork = require( './if-not-fork' );
 const addLabels = require( './tasks/add-labels' );
 const addMilestone = require( './tasks/add-milestone' );
 const assignIssues = require( './tasks/assign-issues' );
@@ -14,6 +11,9 @@ const notifyDesign = require( './tasks/notify-design' );
 const notifyEditorial = require( './tasks/notify-editorial' );
 const triageNewIssues = require( './tasks/triage-new-issues' );
 const wpcomCommitReminder = require( './tasks/wpcom-commit-reminder' );
+const debug = require( './utils/debug' );
+const ifNotClosed = require( './utils/if-not-closed' );
+const ifNotFork = require( './utils/if-not-fork' );
 
 const automations = [
 	{
