+++ projects/github-actions/repo-gardening/src/tasks/assign-issues/index.js
@@ -1,4 +1,4 @@
-const debug = require( '../../debug' );
+const debug = require( '../../utils/debug' );
 
 /* global GitHub, WebhookPayloadPullRequest */
