+++ projects/github-actions/repo-gardening/src/tasks/notify-editorial/index.js
@@ -1,7 +1,7 @@
 const { getInput, setFailed } = require( '@actions/core' );
-const debug = require( '../../debug' );
-const getLabels = require( '../../get-labels' );
-const sendSlackMessage = require( '../../send-slack-message' );
+const debug = require( '../../utils/debug' );
+const getLabels = require( '../../utils/get-labels' );
+const sendSlackMessage = require( '../../utils/send-slack-message' );
 
 /* global GitHub, WebhookPayloadPullRequest */
