+++ projects/github-actions/repo-gardening/src/utils/if-not-closed.js
